fix bug

+++ manager/manager-core/src/main/java/cn/vbill/middleware/porter/manager/service/impl/JobTasksOwnerServiceImpl.java
@@ -104,7 +104,7 @@ public class JobTasksOwnerServiceImpl implements JobTasksOwnerService {
                 if (!controlSettingVo.getToUserIds().isEmpty()) {
                     // 查询预移交者是否为该任务的共享者
                     Integer type = jobTasksOwnerMapper.findOwnerTypeByJobIdAndUserId(controlSettingVo.getJobId(), controlSettingVo.getToUserIds().get(0));
-                    if (type == 2) {
+                    if (type != null && type == 2) {
                         // 提升权限
                         jobTasksOwnerMapper.delete(controlSettingVo.getJobId(), 2, controlSettingVo.getToUserIds().get(0));
                     }
